Implement ESC/VP.net Protocol
- Added `escvpnet` module for ESC/VP.net protocol implementation.
- Created error handling classes for various protocol errors.
- Implemented communication classes for handling ESC/VP21 commands.
- Developed message structure for sending and receiving data.
- Added command headers for password management and projector information.
- Implemented discovery and connection functionalities in the main script.
- Added command-line interface for testing projector connections and commands.
